DRAW-145 refactor: 코드 리뷰 반영
