--- core/src/main/kotlin/com/xorker/draw/mafia/MafiaKeywordUseCase.kt
@@ -1,7 +1,0 @@
-package com.xorker.draw.mafia
-
-import com.xorker.draw.websocket.Session
-
-interface MafiaKeywordUseCase {
-    fun inferAnswer(session: Session, answer: String)
-}
